Using a newer version of angsd and changed -doGlf 3 to 2
I checked and the nInd filter now works with "-doGlf 2" (it didn't work before), so that we can just use "-doGlf 2", so that a separate step to get beagle files will not be needed. In addition, changing -doGlf 3 to 2 seems to avoid a segmentation fault issue that I've been having.